Converted to C#
Tried to keep the logic the same, only changing what was neccessary to go from Java to C#. Didn't have to change that much.

Made stubs for a lot of methods that don't have equivalents in C#.

Project builds correctly now, need to fill in the stubs and it should run.